EDB-319 List Namespace for edgeKV (#13)
* adding command to generate random secret token

* adding create, list and retrieve namespace
